# openapi.yaml
openapi: 3.1.0
info:
  title: Mini Credit Insights API
  description: A backend service for credit insights that ingests CSV bank statements, computes financial insights, and integrates with credit bureaus.
  version: 0.1.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3003
    description: Development server
  - url: https://api.example.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "clx1234567890abcdef"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [USER, ADMIN]
          example: "USER"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "password123"
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "newuser@example.com"
        password:
          type: string
          minLength: 8
          example: "securepassword123"
        role:
          type: string
          enum: [USER, ADMIN]
          default: USER
          example: "USER"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request data"
        details:
          type: array
          items:
            type: object
          example:
            - field: "email"
              message: "Invalid email format"
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      description: Create a new user account (ADMIN only)
      tags: [Authentication]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "newuser@example.com"
              password: "securepassword123"
              role: "USER"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '409':
          description: User already exists
  /auth/login:
    post:
      summary: Login user
      description: Authenticate user and return JWT token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "password123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
...